refactor(park-keeper): 抽離 SettingsTab 為獨立元件
- 自 Home.tsx 純搬移 SettingsTab 設定頁籤至 components/SettingsTab.tsx
- SettingGroup 僅 SettingsTab 使用，一併搬移
- Home.tsx 同步移除未使用 imports（getVersionInfo、LanguageType 與設定頁 lucide icons）

測試：pnpm --filter @app/park-keeper typecheck 與 test --run（31 files / 335 tests）全綠
